# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FOCUS ADHD Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: No profile yet
    put:
      operationId: upsertUserProfile
      tags: [users]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertUserProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: bucket
          in: query
          schema:
            type: string
            enum: [today, later, someday, done]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/steps:
    get:
      operationId: getTaskSteps
      tags: [tasks]
      summary: Get micro-steps for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskStep"
    post:
      operationId: generateTaskSteps
      tags: [tasks]
      summary: AI-generate micro-steps for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskStep"
  /tasks/{taskId}/steps/{stepId}:
    patch:
      operationId: updateTaskStep
      tags: [tasks]
      summary: Update a task step (mark done, skip)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskStepBody"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStep"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nudgeTime:
          type: string
          nullable: true
        struggles:
          type: array
          items:
            type: string
        onboardingComplete:
          type: boolean
        lastOpenAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, onboardingComplete]
    UpsertUserProfileBody:
      type: object
      properties:
        name:
          type: string
        nudgeTime:
          type: string
          nullable: true
        struggles:
          type: array
          items:
            type: string
        onboardingComplete:
          type: boolean
        lastOpenAt:
          type: string
          nullable: true
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        timeEstimateMinutes:
          type: integer
          nullable: true
        bucket:
          type: string
          enum: [today, later, someday, done]
        isFocus:
          type: boolean
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TaskStep"
          nullable: true
      required: [id, title, bucket, isFocus, createdAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        timeEstimateMinutes:
          type: integer
          nullable: true
        bucket:
          type: string
          enum: [today, later, someday, done]
        isFocus:
          type: boolean
      required: [title, bucket]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        timeEstimateMinutes:
          type: integer
          nullable: true
        bucket:
          type: string
          enum: [today, later, someday, done]
        isFocus:
          type: boolean
        completedAt:
          type: string
          nullable: true
    TaskStep:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        title:
          type: string
        timeEstimateMinutes:
          type: integer
        order:
          type: integer
        status:
          type: string
          enum: [pending, done, skipped]
      required: [id, taskId, title, timeEstimateMinutes, order, status]
    UpdateTaskStepBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, done, skipped]
...